pyproject.toml
[project]
name = "timetable-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "axios>=0.4.0",
    "datetime>=6.0",
    "django-cors-headers>=4.9.0",
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "gtfs-realtime-bindings>=1.0.0",
    "pytz>=2025.2",
    "requests>=2.32.5",
]
